Move no-trapping logic to a code_builder module
